trainer/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.9.0
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.5
chromadb>=1.0.0
